+++ src/PaymentGateway.Domain/Services/IBankService.cs
@@ -6,6 +6,6 @@ namespace PaymentGateway.Domain.Services
     public interface IBankService
     {
         Task<PaymentProcessingResult> ProcessPayment(PaymentCardDetails paymentCardDetails,
-            Currency currency, decimal amount);
+            Currency currency, int amount);
     }
 }
